pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "twitter_finetune"
version = "0.0.1"
authors = [{name = "Jan Willem Altink", email = "janwillem@janwillemaltink.eu"}]
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
dependencies = [
    "python-dotenv",
    "openllm",
    "pandas",
]

[project.scripts]
twitter_finetune = "twitter_finetune.main:main"
create_llama_jsonl = "twitter_finetune.utils.dataprep:create_llama_jsonl"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]